requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.4.0
audioread==2.1.9
cachetools==4.2.4
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.1
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
decorator==5.0.9
et-xmlfile==1.1.0
Flask==2.1.1
Flask-WTF==0.15.1
fonttools==4.33.3
ghostscript==0.7
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.4.2
librosa==0.8.1
llvmlite==0.36.0
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.2
numba==0.53.1
numpy==1.21.0
oauthlib==3.1.1
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.0
pandas==1.1.5
pdfminer.six==20211012
pdftopng==0.2.3
pikepdf==4.3.1
Pillow==9.0.0
pluggy==1.0.0
pooch==1.4.0
protobuf==3.19.1
py==1.11.0
py4j==0.10.9
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==4.1.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyspark==3.1.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.8
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.2
six==1.16.0
SoundFile==0.10.3.post1
spotipy==2.19.0
tabulate==0.8.9
threadpoolctl==2.2.0
tomli==2.0.1
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
